Re-organize Notification Behavior samples into sub-folder
Move Header Behaviors doc into root
