+++ components/Behaviors/samples/Notifications/StackedNotificationsBehaviorCustomSample.xaml
@@ -1,5 +1,5 @@
-﻿<!--  Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE file in the project root for more information.  -->
-<Page x:Class="BehaviorsExperiment.Samples.StackedNotificationsBehaviorToolkitSample"
+<!--  Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE file in the project root for more information.  -->
+<Page x:Class="BehaviorsExperiment.Samples.Notifications.StackedNotificationsBehaviorCustomSample"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:behaviors="using:CommunityToolkit.WinUI.Behaviors"
@@ -9,20 +9,19 @@
       xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
       mc:Ignorable="d">
 
-    <StackPanel MinHeight="136"
-                Orientation="Vertical"
-                Spacing="24">
+    <Grid MinHeight="186">
         <Button HorizontalAlignment="Left"
                 VerticalAlignment="Top"
                 Click="Button_Click"
                 Content="Send notification" />
 
-        <muxc:InfoBar HorizontalAlignment="Left"
-                      VerticalAlignment="Bottom"
-                      IsIconVisible="False">
+        <muxc:InfoBar MaxWidth="480"
+                      Margin="24"
+                      HorizontalAlignment="Right"
+                      VerticalAlignment="Bottom">
             <interactivity:Interaction.Behaviors>
                 <behaviors:StackedNotificationsBehavior x:Name="NotificationQueue" />
             </interactivity:Interaction.Behaviors>
         </muxc:InfoBar>
-    </StackPanel>
+    </Grid>
 </Page>
