+++ components/Behaviors/samples/Notifications/StackedNotificationsBehaviorToolkitSample.xaml
@@ -1,5 +1,5 @@
 <!--  Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE file in the project root for more information.  -->
-<Page x:Class="BehaviorsExperiment.Samples.StackedNotificationsBehaviorCustomSample"
+<Page x:Class="BehaviorsExperiment.Samples.Notifications.StackedNotificationsBehaviorToolkitSample"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:behaviors="using:CommunityToolkit.WinUI.Behaviors"
@@ -9,19 +9,20 @@
       xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
       mc:Ignorable="d">
 
-    <Grid MinHeight="186">
+    <StackPanel MinHeight="136"
+                Orientation="Vertical"
+                Spacing="24">
         <Button HorizontalAlignment="Left"
                 VerticalAlignment="Top"
                 Click="Button_Click"
                 Content="Send notification" />
 
-        <muxc:InfoBar MaxWidth="480"
-                      Margin="24"
-                      HorizontalAlignment="Right"
-                      VerticalAlignment="Bottom">
+        <muxc:InfoBar HorizontalAlignment="Left"
+                      VerticalAlignment="Bottom"
+                      IsIconVisible="False">
             <interactivity:Interaction.Behaviors>
                 <behaviors:StackedNotificationsBehavior x:Name="NotificationQueue" />
             </interactivity:Interaction.Behaviors>
         </muxc:InfoBar>
-    </Grid>
+    </StackPanel>
 </Page>
